remove extra api calls and fix reset states on close
